refactor: Coordinator 프로토콜 Util 모듈로 이동
