+++ YakGwa/Core/Common/Sources/Common/Coordinator.swift
@@ -1,19 +1,1 @@
-import UIKit
 
-public protocol Coordinator: AnyObject {
-    var navigationController: UINavigationController? { get }
-    var parentCoordinator: Coordinator? { get set }
-    var childCoordinators: [Coordinator] { get set }
-    
-    func start()
-}
-
-public extension Coordinator {
-    func addChildCoordinator(_ coordinator: Coordinator) {
-        childCoordinators.append(coordinator)
-    }
-    
-    func removeChildCoordinator(_ coordinator: Coordinator) {
-        childCoordinators = childCoordinators.filter { $0 !== coordinator }
-    }
-}
