+++ YakGwa/Core/Util/Sources/Util/Traits/Coordinator.swift
@@ -1,0 +1,26 @@
+//
+//  Coordinator.swift
+//  
+//
+//  Created by Kim Dongjoo on 6/21/24.
+//
+
+import UIKit
+
+public protocol Coordinator: AnyObject {
+    var navigationController: UINavigationController? { get }
+    var parentCoordinator: Coordinator? { get set }
+    var childCoordinators: [Coordinator] { get set }
+    
+    func start()
+}
+
+public extension Coordinator {
+    func addChildCoordinator(_ coordinator: Coordinator) {
+        childCoordinators.append(coordinator)
+    }
+    
+    func removeChildCoordinator(_ coordinator: Coordinator) {
+        childCoordinators = childCoordinators.filter { $0 !== coordinator }
+    }
+}
