+++ YakGwa/Scene/YakgwaDetailScene/Sources/YakgwaDetailScene/Presentation/Calendar/Flows/CalendarVoteCoordinator.swift
@@ -6,7 +6,7 @@
 //
 
 import UIKit
-import Common
+import CoreKit
 import Network
 
 final public class CalendarVoteCoordinator: Coordinator {
