+++ YakGwa/Scene/YakgwaDetailScene/Sources/YakgwaDetailScene/YakgwaAfterSelectionScene/Presentaion/Flows/UserVoteStatusCoordinator.swift
@@ -6,7 +6,7 @@
 //
 
 import UIKit
-import Common
+import CoreKit
 import Network
 
 final public class UserVoteStatusCoordinator: Coordinator {
